Move `assert` methods to `BlackBoxBoundedContext`
