+++ testutil-server/src/test/java/io/spine/testing/server/blackbox/MultitenantBlackBoxContextTest.java
@@ -20,181 +20,19 @@
 
 package io.spine.testing.server.blackbox;
 
+import io.spine.testing.core.given.GivenTenantId;
 import io.spine.testing.server.ShardingReset;
-import io.spine.testing.server.blackbox.command.BbCreateProject;
-import io.spine.testing.server.blackbox.event.BbProjectCreated;
-import io.spine.testing.server.blackbox.event.BbReportCreated;
-import io.spine.testing.server.blackbox.event.BbTaskAdded;
-import io.spine.testing.server.blackbox.event.BbTaskAddedToReport;
-import io.spine.testing.server.blackbox.given.BbProjectRepository;
-import io.spine.testing.server.blackbox.given.BbProjectViewRepository;
-import io.spine.testing.server.blackbox.given.BbReportRepository;
-import io.spine.testing.server.blackbox.given.RepositoryThrowingExceptionOnClose;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-
-import static com.google.common.collect.ImmutableSet.of;
-import static io.spine.testing.client.blackbox.Count.count;
-import static io.spine.testing.client.blackbox.Count.once;
-import static io.spine.testing.client.blackbox.Count.thrice;
-import static io.spine.testing.client.blackbox.Count.twice;
-import static io.spine.testing.client.blackbox.VerifyAcknowledgements.acked;
-import static io.spine.testing.core.given.GivenTenantId.newUuid;
-import static io.spine.testing.server.blackbox.given.Given.addTask;
-import static io.spine.testing.server.blackbox.given.Given.createProject;
-import static io.spine.testing.server.blackbox.given.Given.createReport;
-import static io.spine.testing.server.blackbox.given.Given.newProjectId;
-import static io.spine.testing.server.blackbox.given.Given.taskAdded;
-import static io.spine.testing.server.blackbox.verify.state.VerifyState.exactly;
-import static io.spine.testing.server.blackbox.verify.state.VerifyState.exactlyOne;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @ExtendWith(ShardingReset.class)
 @DisplayName("Multi tenant Black Box Bounded Context should")
-class MultitenantBlackBoxContextTest {
+class MultitenantBlackBoxContextTest
+        extends BlackBoxBoundedContextTest<MultitenantBlackBoxContext> {
 
-    private MultitenantBlackBoxContext projects;
-
-    @BeforeEach
-    void setUp() {
-        projects = BlackBoxBoundedContext.multitenant()
-                                         .with(new BbProjectRepository(),
-                                               new BbProjectViewRepository());
-    }
-
-    @AfterEach
-    void tearDown() {
-        projects.close();
-    }
-
-    @Nested
-    @DisplayName("verify state of")
-    class VerifyStateOf {
-
-        @Test
-        @DisplayName("a single aggregate")
-        void aggregate() {
-            BbCreateProject createProject = createProject();
-            BbProject expectedProject = createdProjectAggregate(createProject);
-            projects.withTenant(newUuid())
-                    .receivesCommand(createProject)
-                    .assertThat(exactlyOne(expectedProject));
-        }
-
-        @Test
-        @DisplayName("several aggregates")
-        void aggregates() {
-            BbCreateProject createProject1 = createProject();
-            BbCreateProject createProject2 = createProject();
-            BbProject expectedProject1 = createdProjectAggregate(createProject1);
-            BbProject expectedProject2 = createdProjectAggregate(createProject2);
-            projects.withTenant(newUuid())
-                    .receivesCommands(createProject1, createProject2)
-                    .assertThat(exactly(BbProject.class, of(expectedProject1, expectedProject2)));
-        }
-
-        @Test
-        @DisplayName("a single projection")
-        void projection() {
-            BbCreateProject createProject = createProject();
-            BbProjectView expectedProject = createProjectView(createProject);
-            projects.withTenant(newUuid())
-                    .receivesCommand(createProject)
-                    .assertThat(exactlyOne(expectedProject));
-        }
-
-        @Test
-        @DisplayName("several projections")
-        void projections() {
-            BbCreateProject createProject1 = createProject();
-            BbCreateProject createProject2 = createProject();
-            BbProjectView expectedProject1 = createProjectView(createProject1);
-            BbProjectView expectedProject2 = createProjectView(createProject2);
-            projects.withTenant(newUuid())
-                    .receivesCommands(createProject1, createProject2)
-                    .assertThat(exactly(BbProjectView.class,
-                                        of(expectedProject1, expectedProject2)));
-        }
-
-        private BbProject createdProjectAggregate(BbCreateProject createProject) {
-            return BbProjectVBuilder.newBuilder()
-                                    .setId(createProject.getProjectId())
-                                    .build();
-        }
-
-        private BbProjectView createProjectView(BbCreateProject createProject) {
-            return BbProjectViewVBuilder.newBuilder()
-                                        .setId(createProject.getProjectId())
-                                        .build();
-        }
-    }
-
-    @Test
-    @DisplayName("receive and handle a single command")
-    void receivesACommand() {
-        projects.withTenant(newUuid())
-                .receivesCommand(createProject())
-                .assertThat(acked(once()).withoutErrorsOrRejections())
-                .assertThat(VerifyEvents.emittedEvent(BbProjectCreated.class, once()));
-    }
-
-    @Test
-    @DisplayName("receive and handle multiple commands")
-    void receivesCommands() {
-        BbProjectId projectId = newProjectId();
-        projects.withTenant(newUuid())
-                .receivesCommand(createProject(projectId))
-                .receivesCommands(addTask(projectId), addTask(projectId), addTask(projectId))
-                .assertThat(acked(count(4)).withoutErrorsOrRejections())
-                .assertThat(VerifyEvents.emittedEvent(count(4)))
-                .assertThat(VerifyEvents.emittedEvent(BbProjectCreated.class, once()))
-                .assertThat(VerifyEvents.emittedEvent(BbTaskAdded.class, thrice()));
-    }
-
-    @Test
-    @DisplayName("receive and react on single event")
-    void receivesEvent() {
-        BbProjectId projectId = newProjectId();
-        projects.with(new BbReportRepository())
-                .withTenant(newUuid())
-                .receivesCommand(createReport(projectId))
-                .receivesEvent(taskAdded(projectId))
-                .assertThat(acked(twice()).withoutErrorsOrRejections())
-                .assertThat(VerifyEvents.emittedEvent(thrice()))
-                .assertThat(VerifyEvents.emittedEvent(BbReportCreated.class, once()))
-                .assertThat(VerifyEvents.emittedEvent(BbTaskAddedToReport.class, once()));
-    }
-
-    @Test
-    @DisplayName("receive and react on multiple events")
-    void receivesEvents() {
-        BbProjectId projectId = newProjectId();
-        projects.with(new BbReportRepository())
-                .withTenant(newUuid())
-                .receivesCommand(createReport(projectId))
-                .receivesEvents(taskAdded(projectId), taskAdded(projectId), taskAdded(projectId))
-                .assertThat(acked(count(4)).withoutErrorsOrRejections())
-                .assertThat(VerifyEvents.emittedEvent(count(7)))
-                .assertThat(VerifyEvents.emittedEvent(BbReportCreated.class, once()))
-                .assertThat(VerifyEvents.emittedEvent(BbTaskAddedToReport.class, thrice()));
-    }
-
-    @Test
-    @DisplayName("throw Illegal State Exception on Bounded Context close error")
-    void throwIllegalStateExceptionOnClose() {
-        assertThrows(IllegalStateException.class, () ->
-                BlackBoxBoundedContext
-                        .multitenant()
-                        .with(new RepositoryThrowingExceptionOnClose() {
-                            @Override
-                            protected void throwException() {
-                                throw new RuntimeException("Expected error");
-                            }
-                        })
-                        .close());
+    @Override
+    BlackBoxBoundedContext<MultitenantBlackBoxContext> newInstance() {
+        return BlackBoxBoundedContext.multitenant()
+                                     .withTenant(GivenTenantId.newUuid());
     }
 }
